A sanitizer build that can reach a level, and five undefined shifts
`timasan` is built from the shipping `main.c`, and a DOS game's entry point has
no command line - so it can be pointed at the intro and nothing else. Every
frame in `parts.c` and `machine.c` belongs to code a *level* reaches, which
needs `--restore`, which only `devmain.c` has. `make asandev` builds that one.
Four levels and every solution snapshot tried so far are **clean under
AddressSanitizer**; the intro already was.

Then the UBSan half, which is mostly not a defect list - the guest's records
are packed and reached at odd offsets by design, so "load of misaligned
address" is the model working. Setting those aside left three reports about
*shifts*, and chasing the family found two more:

  - `machine.c`'s `PART(obj).fy <<= 9` on -1000 and `trig.c`'s
    `(int16_t)(r - 0x400) << 4` shift a negative value. The original's `shl`
    has no sign in it; C calls this undefined. Shifting the `uint16_t` is the
    same bits.
  - `((int32_t)(uint16_t)x << 16)` in `parts.c` and twice in `machine.c`
    overflows a **signed** int whenever x is above 0x7fff - and signed
    overflow is one of the undefined behaviours a compiler will actually act
    on, unlike the alignment ones. Twenty other sites in the port build the
    same 32-bit value as `((uint32_t)x << 16)` with the cast on the result.
    Three typos in a family of twenty-three, with the majority spelling right.

None of the five changes a computed value on any machine this runs on; they
stop the port relying on the compiler's goodwill.

Verified: `verify.py --all` left STATUS.md untouched, the intro is 66 flips
byte for byte, the briefing 0 of 307200 pixels, `check_sound` 55 runs
identical, and 33 of 33 solutions solve - which is the check that matters for
`trig.c` and the momentum clamp.

+++ reconstruct/src/parts.c
@@ -4342,7 +4342,7 @@ uint16_t part_step_1a82(uint16_t part)
         mass = PARTKIND((int16_t)PART(si).kind).weight;
 
         v04 = (int16_t)long_divide(
-            ((int32_t)(uint16_t)v08 << 16) | (uint16_t)v0a,
+            (int32_t)(((uint32_t)(uint16_t)v08 << 16) | (uint16_t)v0a),
             (int32_t)mass);
 
         PART(si).vel_x =
